backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0
httpx==0.26.0
tenacity==8.2.3
pytest==7.4.4
pytest-asyncio==0.23.3
hypothesis==6.98.3
pymongo==4.6.1
pymongo==4.6.1
